requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
jinja2>=3.1.0
aiofiles>=23.0.0

# Database
sqlmodel>=0.0.14

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Store for RAG
chromadb>=0.4.0

# PDF Processing
pypdf>=3.17.0

# LLM Providers
openai>=1.3.0
groq>=0.4.0

# Embeddings
sentence-transformers>=2.2.0

# Utilities
httpx==0.25.2
